backend/requirements.txt
asgiref==3.9.2
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
sqlparse==0.5.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==3.0.1